refactor(examples): Move concept files into examples
