+++ examples/Domain/Store/Documents/Product.php
@@ -7,12 +7,12 @@
 
 declare(strict_types=1);
 
-namespace ArtoxLab\DStore\Tests\Domain\Store\Documents;
+namespace ArtoxLab\DStore\Examples\Domain\Store\Documents;
 
 
 use ArtoxLab\DStore\Interfaces\DocumentInterface;
-use ArtoxLab\DStore\Tests\Domain\Entities\Product as ProductEntity;
-use ArtoxLab\DStore\Tests\Domain\Store\Documents\Product\IndexByBrandId;
+use ArtoxLab\DStore\Examples\Domain\Entities\Product as ProductEntity;
+use ArtoxLab\DStore\Examples\Domain\Store\Documents\Product\IndexByBrandId;
 use ArtoxLab\Entities\States\StateItem;
 
 class Product implements DocumentInterface
