+++ examples/Domain/Store/Documents/ProductEntityBuilder.php
@@ -7,11 +7,11 @@
 
 declare(strict_types=1);
 
-namespace ArtoxLab\DStore\Tests\Domain\Store\Documents;
+namespace ArtoxLab\DStore\Examples\Domain\Store\Documents;
 
-use ArtoxLab\DStore\Tests\Domain\Entities\Brand;
-use ArtoxLab\DStore\Tests\Domain\Entities\BrandScore;
-use ArtoxLab\DStore\Tests\Domain\Entities\Product as ProductEntity;
+use ArtoxLab\DStore\Examples\Domain\Entities\Brand;
+use ArtoxLab\DStore\Examples\Domain\Entities\BrandScore;
+use ArtoxLab\DStore\Examples\Domain\Entities\Product as ProductEntity;
 use ArtoxLab\Entities\States\StateItem;
 
 class ProductEntityBuilder
